pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Flashcard App API with OTP verification and AI-powered content generation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.8.1",
    "apscheduler>=3.11.0",
    "bcrypt==3.2.0",
    "easyocr>=1.7.0",
    "email-validator>=2.1.0",
    "fastapi>=0.116.1",
    "fastapi-mail>=1.4.1",
    "groq>=0.31.1",
    "ipykernel>=6.30.1",
    "mercadopago>=2.3.0",
    "openai>=1.107.1",
    "passlib[bcrypt]>=1.7.4",
    "prisma>=0.15.0",
    "pymupdf>=1.26.4",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "pytz>=2025.2",
    "redis>=6.4.0",
    "tiktoken>=0.11.0",
    "uvicorn>=0.35.0",
]
